grep: Use extended regex instead of perl regex
At least with GNU grep, PCRE are much slower than extended regexps
even for simple patterns.

A disadvantage of extended regexps here is that micro itself uses PCRE
for its searches. (In particular, this inconsistency may cause wrong
selection in selectAtCurLine().) However, in this regard the plugin is
far from perfect anyway, so for now speed is more important.